Use multipage plotly. Add screen/ static dashboard
